requirements.txt
absl-py==1.4.0
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
imageio==2.27.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-loader==0.2
MarkupSafe==2.1.2
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tifffile==2023.3.21
tomli==2.0.1
tzdata==2023.3
Werkzeug==2.2.3
zipp==3.15.0
